# openapi/blerp.v1.yaml
openapi: 3.1.0
info:
  title: Blerp Identity Service API
  version: 1.0.0-draft
  contact:
    name: Blerp Support
    url: https://github.com/e6qu/blerp
    email: adi11235@gmail.com
  summary: REST + JSON interface for Blerp Identity Service
  description: >
    Clean-room OpenAPI contract describing the public REST API for user, session,
    organization, and webhook management. This specification is the authoritative
    contract for backend/frontend work and must exist locally before any detailed
    implementation plan is executed.
...
servers:
  - url: https://api.blerp.dev
    description: Production
  - url: https://staging-api.blerp.dev
    description: Staging
  - url: http://localhost:4000
    description: Local development (docker-compose)
security:
  - SecretKey: []
...
components:
  securitySchemes:
    SecretKey:
      type: http
      scheme: bearer
      bearerFormat: secret-key
      description: Server-to-server secret API key (sk_live_* / sk_test_*).
    PublishableKey:
      type: http
      scheme: bearer
      bearerFormat: publishable-key
      description: Client-side publishable API key (pk_live_* / pk_test_*).
    SessionToken:
      type: http
      scheme: bearer
      bearerFormat: session-token
      description: Session token returned after sign-in/up (`sess_*`).
  schemas:
    ErrorResponse:
      type: object
      # BUG-47/57: blerp emits the Clerk-compatible `errors[]` envelope
      # (canonical for new SDK code) from any path that throws a
      # `BlerpError`. Many hand-rolled `res.status(N).json({ error: ... })`
      # paths in controller catch blocks still emit only the singular
      # `error` field — so `error` is required while `errors` stays
      # optional until those sites migrate to throwing `BlerpError`
      # (tracked in DO_NEXT.md). Generated SDK clients should read
      # `errors[0]` when present and fall back to `error` otherwise.
      # BUG-60 (codex r4): had `errors` required here; SDK consumers
      # would crash on legacy-shape 400s.
      required: [error]
      properties:
        errors:
          type: array
          minItems: 1
          items:
            type: object
            required: [code, message, long_message]
            properties:
              code:
                type: string
              message:
                type: string
              long_message:
                type: string
              meta:
                type: object
                additionalProperties: true
        error:
          type: object
          required: [code, message]
          description: Legacy singular alias; prefer `errors[0]` in new code.
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
            doc_url:
              type: string
              format: uri
            request_id:
              type: string
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          minimum: 0
        next_cursor:
          type: string
        previous_cursor:
          type: string
    EmailAddress:
      type: object
      required: [id, email, verification]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        verification:
          type: object
          required: [status]
          properties:
            status:
              type: string
              enum: [unverified, pending, verified, failed]
            strategy:
              type: string
              enum: [email_code, email_link]
            attempts:
              type: integer
              minimum: 0
            last_attempted_at:
              type: string
              format: date-time
    PhoneNumber:
      type: object
      required: [id, number, verification]
      properties:
        id:
          type: string
          format: uuid
        number:
          type: string
        verification:
          $ref: "#/components/schemas/EmailAddress/properties/verification"
    PasskeyCredential:
      type: object
      required: [id, friendly_name]
      properties:
        id:
          type: string
        friendly_name:
          type: string
        transports:
          type: array
          items:
            type: string
          description: WebAuthn transports captured at registration. Always present, may be empty until the registration verifier surfaces them.
        created_at:
          type: [string, "null"]
          format: date-time
        last_used_at:
          type: [string, "null"]
          format: date-time
    PublicMetadata:
      type: object
      additionalProperties: true
    PrivateMetadata:
      type: object
      additionalProperties: true
    UsageResponse:
      type: object
      required: [users, organizations, sessions, limits]
      properties:
        users:
          type: integer
        organizations:
          type: integer
        sessions:
          type: integer
        limits:
          type: object
          required: [maxUsers, maxOrganizations, maxSessions]
          properties:
            maxUsers:
              type: integer
            maxOrganizations:
              type: integer
            maxSessions:
              type: integer
    User:
      type: object
      required:
        - id
        - status
        - created_at
        - email_addresses
      properties:
        id:
          type: string
          format: uuid
        external_id:
          type: string
        username:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        image_url:
          type: string
          format: uri
        has_image:
          type: boolean
        primary_email_id:
          type: string
          format: uuid
        primary_phone_id:
          type: string
          format: uuid
        email_addresses:
          type: array
          items:
            $ref: "#/components/schemas/EmailAddress"
        phone_numbers:
          type: array
          items:
            $ref: "#/components/schemas/PhoneNumber"
        passkeys:
          type: array
          items:
            $ref: "#/components/schemas/PasskeyCredential"
        public_metadata:
          $ref: "#/components/schemas/PublicMetadata"
        private_metadata:
          $ref: "#/components/schemas/PrivateMetadata"
        unsafe_metadata:
          $ref: "#/components/schemas/PublicMetadata"
        status:
          type: string
          enum: [active, inactive, banned]
        locked:
          type: boolean
        password_enabled:
          type: boolean
          description: |
            Whether the user has a password set. Matches Clerk's
            `passwordEnabled` field (BUG-123 codex r22). Dashboard /
            SDK UIs use this to decide whether to render a "Set
            password" CTA.
        totp_enabled:
          type: boolean
          description: Whether TOTP (authenticator app) is enabled for this user.
        backup_code_enabled:
          type: boolean
          description: Whether backup codes are available for account recovery.
        two_factor_enabled:
          type: boolean
          description: Whether any form of two-factor authentication is enabled.
        last_sign_in_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
...
    Organization:
      type: object
      required: [id, project_id, name]
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        public_metadata:
          type: object
          additionalProperties: true
        private_metadata:
          type: object
          additionalProperties: true
        domains:
          type: array
          items:
            type: object
            properties:
              domain:
                type: string
              status:
                type: string
                enum: [pending, verified, rejected]
              verified_at:
                type: string
                format: date-time
        settings:
          type: object
          properties:
            allow_invites:
              type: boolean
            default_role:
              type: string
            mfa_policy:
              type: string
              enum: [optional, required, disabled]
        created_at:
          type: string
          format: date-time
...
    Signup:
      type: object
      required: [id, strategy, status]
      properties:
        id:
          type: string
        identifier:
          type: string
        strategy:
          type: string
          enum: [password, magic_link, otp, passkey, oauth]
        status:
          type: string
          enum: [needs_identifier, needs_verification, complete, abandoned]
        verification:
          type: object
          properties:
            channel:
              type: string
              enum: [email_code, email_link, sms]
            code:
              type: string
            expires_at:
              type: string
              format: date-time
        created_session_id:
          type: string
        expires_at:
          type: string
          format: date-time
...
    AuditLogEntry:
      type: object
      required: [id, project_id, actor, action, resource, created_at]
      properties:
        id:
          type: string
        project_id:
          type: string
        actor:
          type: object
          required: [type, id]
          properties:
            type:
              type: string
              enum: [user, api_key, system]
            id:
              type: string
            email:
              type: string
              format: email
        action:
          type: string
        resource:
          type: object
          properties:
            type:
              type: string
            id:
              type: string
            display:
              type: string
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
...
paths:
...
  /v1/auth/signups:
    post:
      tags: [Auth]
      summary: Create pending signup
      description: Creates a signup intent and dispatches verification challenges for a new end user.
      operationId: createSignup
      security:
        - SecretKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [strategy]
              properties:
                email:
                  type: string
                  format: email
                phone:
                  type: string
                password:
                  type: string
                strategy:
                  type: string
                  enum: [password, magic_link, otp, passkey, oauth]
      responses:
        "201":
          description: Signup created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Signup"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/users/{user_id}/metadata:
    patch:
      tags: [Users]
      summary: Update user metadata
      description: Updates the public, private, or unsafe metadata for a user.
      operationId: updateUserMetadata
      security:
        - SecretKey: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                public_metadata:
                  type: object
                  additionalProperties: true
                private_metadata:
                  type: object
                  additionalProperties: true
                unsafe_metadata:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /v1/organizations/{organization_id}/metadata:
    patch:
      tags: [Organizations]
      summary: Update organization metadata
      description: Updates the public or private metadata for an organization.
      operationId: updateOrganizationMetadata
      security:
        - SecretKey: []
      parameters:
        - name: organization_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                public_metadata:
                  type: object
                  additionalProperties: true
                private_metadata:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: Updated organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
...
  /v1/audit_logs:
    get:
      tags: [Audit]
      summary: List audit log entries
      description: Returns paginated audit log entries for compliance and monitoring.
      operationId: listAuditLogs
      security:
        - SecretKey: []
      parameters:
        - in: query
          name: actor_id
          schema:
            type: string
        - in: query
          name: resource_type
          schema:
            type: string
        - in: query
          name: resource_id
          schema:
            type: string
        - in: query
          name: created_before
          schema:
            type: string
            format: date-time
        - in: query
          name: created_after
          schema:
            type: string
            format: date-time
        - in: query
          name: cursor
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Audit log entries
          content:
            application/json:
              schema:
                type: object
                required: [data, total_count]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AuditLogEntry"
                  # BUG-48/58: Clerk-canonical pagination field. SDK
                  # clients should read total_count; meta is a legacy
                  # alias kept for one release.
                  total_count:
                    type: integer
                    minimum: 0
                  meta:
                    $ref: "#/components/schemas/PaginationMeta"
...
  /v1/usage:
    get:
      tags: [System]
      summary: Get usage statistics
      description: Retrieve resource usage and quota consumption for the tenant.
      operationId: getUsage
      security:
        - SecretKey: []
      responses:
        "200":
          description: Usage stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageResponse"
...